I forgot the 'double' column in a migration

+++ db/migrate/20140624145055_create_no_cms_events_locations.rb
@@ -3,8 +3,8 @@ class CreateNoCmsEventsLocations < ActiveRecord::Migration
     create_table :no_cms_events_locations do |t|
       t.string :name
       t.text :description
-      t.double :lat
-      t.double :lng
+      t.float :lat
+      t.float :lng
 
       t.timestamps
     end
